+++ pytation/__init__.py
@@ -14,8 +14,9 @@
 
 from .version import *
 from .context import Context
+from .analysis import AnalysisContext
 from .api import declare_test
 
-__all__ = ['Context', 'declare_test',
+__all__ = ['Context', 'AnalysisContext', 'declare_test',
            '__version__', '__title__', '__description__', '__url__',
            '__author__', '__author_email__', '__license__', '__copyright__']