+++ pytation/api.py
@@ -14,8 +14,8 @@
 
 """The API for packages using this manufacturing test framework."""
 
-
-from . import Context
+from pytation.context import Context
+from pytation.analysis import AnalysisContext
 
 
 def declare_test(devices: list[str] = None):
@@ -23,7 +23,7 @@ def declare_test(devices: list[str] = None):
 
     :param devices: The list of expected devices, using the same
         device names used in the station definition.
-    :return: The decorated test function(context, config).
+    :return: The decorated test function(context).
     """
     def decorator_repeat(func):
         if devices:
@@ -32,28 +32,53 @@ def declare_test(devices: list[str] = None):
     return decorator_repeat
 
 
-def test_prototype(context: Context, config: dict[str, object]):
+def test_prototype(context: Context):
     """The prototype for a test function.
 
     :param context: The pytation test station context.
-    :param config: Optional test configuration options.
+        Interesting members include:
+            - state: The current state name, which must match a state in
+              the station configuration.
+            - env: The station environment.
+            - fs: The filesystem instance for use by the test.
+            - config: The dict[str, object] of test configuration options.
+              The test may modify this configuration in place, and the
+              station will store the modified version for future analysis.
+
+        Interesting methods include:
+            - expand_str
+            - path
+            - section, section_enter, section_exit
+            - progress
+            - wait_for_user
+            - prompt
+
     :return: One of the following:
         * None: test passed
-        * result: integer return code with 0=success, 1=fail
+        * result: integer return code with 0=success, anything else=fail
         * result, details: The integer return code along with a dict
           of support details that will be logged.  The details are
-          also added to the context.
+          also added to the context.  details must be JSON serializable.
+    :raise Exception: Test fails.
 
     A module with a "run" function that conforms to this prototype
     may also be used as a test.  The module may also provide an
-    "analyze(context)" function for use with the analyze entry point.
+    "analysis" function conforming to the :func:`analysis_prototype`.
     """
     return 0, {}
-    result = fn(self, config)
-    if result is None:
-        result = 0
-    elif not isinstance(result, int):
-        result, detail = result
+
+
+def analysis_prototype(context: AnalysisContext):
+    """The prototype for an analysis function.
+
+    :param context: The pytation test station analysis context.
+    :return: 0 or error code.
+    :raise Exception: On analysis failure.
+
+    A module with a "run" function may also contain an "analysis" function
+    that conforms to this prototype.
+    """
+    return 0
 
 
 class Device:
@@ -62,11 +87,10 @@ class Device:
     NAME = ''
     """The user-meaningful, descriptive test name"""
 
-    def setup(self, context: Context, config: dict[str, object]):
+    def setup(self, context: Context):
         """Open and initialize the device.
 
         :param context: The test station context.
-        :param config: Optional additional device configuration options.
         """
         raise NotImplementedError("Device.setup")